+++ examples/resources/voipms_did/resource.tf
@@ -1,9 +1,3 @@
-locals {
-  my_pops = {
-    ny7 = "newyork7.voip.ms"
-  }
-}
-
 data "voipms_voicemail" "johns_voicemail" {
   name = "John"
 }
@@ -16,16 +10,20 @@ data "voipms_subaccount" "gateway" {
   username = "gateway"
 }
 
+data "voipms_server" "nyc" {
+  hostname = "newyork7.voip.ms"
+}
+
 resource "voipms_did" "home" {
   did                  = "5550001001"
   note                 = "Home line"
-  routing              = "account:${data.voipms_subaccount.gateway.account}"
-  pop_hostname         = local.my_pops.ny7
+  routing              = data.voipms_subaccount.gateway.route
+  pop_hostname         = data.voipms_server.nyc.hostname
   dialtime             = 30
-  voicemail_name       = data.voipms_voicemail.johns_voicemail.name
-  failover_busy        = "vm:${data.voipms_voicemail.johns_voicemail.name}"
-  failover_noanswer    = "vm:${data.voipms_voicemail.johns_voicemail.name}"
-  failover_unreachable = "fwd:${data.voipms_forwarding.mobile.description}"
+  voicemail            = data.voipms_voicemail.johns_voicemail.id
+  failover_busy        = data.voipms_voicemail.johns_voicemail.route
+  failover_noanswer    = data.voipms_voicemail.johns_voicemail.route
+  failover_unreachable = data.voipms_forwarding.mobile.route
 
   sms_enabled     = true
   webhook         = "https://example.lambda-url.us-east-1.on.aws/"